Refactor : Game 페이지의 index에서 사용되는 3개의 useState를 userReducer로 변경
상세:
- useState 사용된 3개의 상태를 Game 페이지용 useReducer로 통합변경
- useReducer에 사용되는 action과 type들은 기존의 global의 것들과 통합
- 기존의 global에서 사용되는 type들과 action들을 변경없이 이용할 수 있게끔 해주는
  방식으로 통합함

#488
