+++ client/src/presentation/pages/Game/store/index.js
@@ -1,0 +1,2 @@
+export { default as gameReducer } from './gameReducer';
+export { default as gameState } from './gameState';
